pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 70.1"]

[project]
dependencies = [
  "fastapi[standard]==0.134.0",
  "pydantic-settings==2.13.1",
  "sqlalchemy==2.0.46",
]
description = "General Purpose ETL"
dynamic = ["version"]
license = {file = "LICENSE"}
name = "aisomething"
readme = "README.md"

[tool.setuptools.dynamic]
version = {attr = "aisomething.__version__"}

[project.optional-dependencies]
build = ["build"]
format = [
  "black>24",
]
lint = [
  "ruff>=0.5",
  "mypy>=1.10.1",
  "tox>=4",
]
test = [
  "pytest>=8",
  "pytest-cov",
  "pytest-sugar",
  "coverage[toml]",
]

[tool.setuptools]
packages = ["aisomething"]

[tool.setuptools_scm]

[tool.ruff]
extend-exclude = [
  "conf.py",
]
line-length = 120
lint.ignore = [
  "D",
  "COM812", # Conflicts with the formatter
  "ISC001", # Conflicts with the formatter  # "ANN101", # "missing-type-self"
  "PT001", # https://github.com/astral-sh/ruff/issues/8796#issuecomment-1825907715
  "PT023", # https://github.com/astral-sh/ruff/issues/8796#issuecomment-1825907715
]
lint.select = ["ALL"]
src = ["src"]
target-version = "py311"

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = [
  # "F401", # Imported but unused
  # "F403", # Wildcard imports
]
"docs/**" = [
  # "INP001", # Requires __init__.py but docs folder is not a package.
]
"tests/**" = [
  "S101", # Use of `assert` detected
  "D103", # Missing docstring in public function
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`(https://github.com/astral-sh/ruff/issues/5434)
keep-runtime-typing = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
disallow_untyped_defs = true # Functions need to be annotated
exclude = [
  "my-project-\\d+", # Ignore temporary folder created by setuptools when building an sdist
  "venv.*/",
  "build/",
  "dist/",
] 
warn_unused_ignores = true
